requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
docopt==0.6.2
fastapi==0.115.12
Flask==3.1.0
fonttools==4.56.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mqtt-client==1.6.1
numpy==2.2.4
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
terminaltables==3.1.10
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
Werkzeug==3.1.3
